requirements.txt
python-dotenv>=1.0.0
spotipy>=2.23.0
numpy>=1.24.3
sentence-transformers>=2.2.2
tqdm>=4.65.0
tabulate>=0.9.0 